+++ tools/nodejs_api/src_cpp/all_each_async_worker.h
@@ -4,16 +4,17 @@
 
 using namespace Napi;
 
-class AllAsyncWorker : public AsyncWorker {
+class AllEachAsyncWorker : public AsyncWorker {
 
 public:
-    AllAsyncWorker(Function& callback, shared_ptr<kuzu::main::QueryResult>& queryResult);
-    virtual ~AllAsyncWorker() {};
+    AllEachAsyncWorker(Function& callback, shared_ptr<kuzu::main::QueryResult>& queryResult, std::string executionType);
+    virtual ~AllEachAsyncWorker() {};
 
     void Execute();
     void OnOK();
 
 private:
+    std::string executionType;
     shared_ptr<kuzu::main::QueryResult> queryResult;
     std::vector<std::vector<std::string>> allResult;
 };
