+++ tools/nodejs_api/src_cpp/execute_async_worker.cpp
@@ -1,6 +1,8 @@
-#include "ExecuteAsyncWorker.h"
+#include "execute_async_worker.h"
+
 #include <chrono>
 #include <thread>
+
 #include "node_query_result.h"
 
 ExecuteAsyncWorker::ExecuteAsyncWorker(Function& callback, shared_ptr<kuzu::main::Connection>& connection,
