+++ tools/nodejs_api/src_cpp/node_query_result.cpp
@@ -1,7 +1,6 @@
 #include "node_query_result.h"
-#include "AllAsyncWorker.h"
-#include "EachAsyncWorker.h"
 
+#include "all_each_async_worker.h"
 #include "main/kuzu.h"
 
 using namespace kuzu::main;
@@ -11,8 +10,6 @@ Napi::FunctionReference NodeQueryResult::constructor;
 Napi::Object NodeQueryResult::Wrap(Napi::Env env, shared_ptr<kuzu::main::QueryResult>& queryResult) {
     Napi::HandleScope scope(env);
     Napi::Object obj = DefineClass(env, "NodeQueryResult", {
-      InstanceMethod("hasNext", &NodeQueryResult::HasNext),
-      InstanceMethod("getNext", &NodeQueryResult::GetNext),
       InstanceMethod("close", &NodeQueryResult::Close),
       InstanceMethod("all", &NodeQueryResult::All),
       InstanceMethod("each", &NodeQueryResult::Each),
@@ -23,35 +20,6 @@ Napi::Object NodeQueryResult::Wrap(Napi::Env env, shared_ptr<kuzu::main::QueryRe
 }
 
 NodeQueryResult::NodeQueryResult(const Napi::CallbackInfo& info) : Napi::ObjectWrap<NodeQueryResult>(info)  {}
-
-Napi::Value NodeQueryResult::HasNext(const Napi::CallbackInfo& info) {
-    Napi::Env env = info.Env();
-    Napi::HandleScope scope(env);
-
-    if (info.Length()!=0) {
-        Napi::TypeError::New(env, "hasNext takes no arguments").ThrowAsJavaScriptException();
-        return Napi::Object::New(env);
-    }
-
-    return Napi::Boolean::New(env, this->queryResult->hasNext());
-}
-
-Napi::Value NodeQueryResult::GetNext(const Napi::CallbackInfo& info) {
-    Napi::Env env = info.Env();
-    Napi::HandleScope scope(env);
-
-    if (info.Length()!=0) {
-        Napi::TypeError::New(env, "getNext takes no arguments").ThrowAsJavaScriptException();
-        return Napi::Object::New(env);
-    }
-
-    auto row = this->queryResult->getNext();
-    Napi::Array obj = Napi::Array::New(env, row->len());
-    for (int j=0; j < row->len(); j++) {
-        obj.Set(j, row->getValue(j)->toString());
-    }
-   return obj;
-}
 
 void NodeQueryResult::Close(const Napi::CallbackInfo& info) {
     Napi::Env env = info.Env();
@@ -71,7 +39,7 @@ Napi::Value NodeQueryResult::All(const Napi::CallbackInfo& info) {
 
     Function callback = info[0].As<Function>();
 
-    AllAsyncWorker * asyncWorker = new AllAsyncWorker(callback, queryResult);
+    AllEachAsyncWorker * asyncWorker = new AllEachAsyncWorker(callback, queryResult, "all");
     asyncWorker->Queue();
     return info.Env().Undefined();
 }
@@ -87,7 +55,7 @@ Napi::Value NodeQueryResult::Each(const Napi::CallbackInfo& info) {
 
     Function callback = info[0].As<Function>();
 
-    EachAsyncWorker * asyncWorker = new EachAsyncWorker(callback, queryResult);
+    AllEachAsyncWorker * asyncWorker = new AllEachAsyncWorker(callback, queryResult, "each");
     asyncWorker->Queue();
     return info.Env().Undefined();
 }
