+++ tools/nodejs_api/src_nodejs/common.js
@@ -1,0 +1,10 @@
+function callbackWrapper(err, callback) {
+    if (err){
+        console.log("There is an error!!!" + err);
+        throw err;
+    } else {
+        callback()
+    }
+}
+module.exports = callbackWrapper
+
